preparation/requirements.txt
tqdm
opencv-python
ffmpeg-python
sentencepiece
scikit-image
librosa
